refactor(BuySell): split PaymentMethod into type and account (work in progress)
