+++ src/exchange/exchange.js
@@ -2,7 +2,7 @@ var assert = require('assert');
 var Helpers = require('./helpers');
 
 class Exchange {
-  constructor (delegate, TradeClass, QuoteClass, PaymentMethodClass) {
+  constructor (delegate, TradeClass, QuoteClass, PaymentMediumClass) {
     assert(delegate, 'ExchangeDelegate required');
     assert(TradeClass, 'Trade class required');
     assert(QuoteClass, 'Quote class required');
@@ -12,7 +12,7 @@ class Exchange {
     this._trades = [];
     this._TradeClass = TradeClass;
     this._QuoteClass = QuoteClass;
-    this._PaymentMethodClass = PaymentMethodClass;
+    this._PaymentMediumClass = PaymentMediumClass;
   }
 
   get debug () { return this._debug; }
@@ -41,11 +41,11 @@ class Exchange {
   get delegate () { return this._delegate; }
 
   getBuyMethods () {
-    return this._PaymentMethodClass.fetchAll(undefined, 'BTC', this._api);
+    return this._PaymentMediumClass.getAll(undefined, 'BTC', this._api);
   }
 
   getSellMethods () {
-    return this._PaymentMethodClass.fetchAll('BTC', undefined, this._api);
+    return this._PaymentMediumClass.getAll('BTC', undefined, this._api);
   }
 
   getBuyQuote (amount, baseCurrency, quoteCurrency) {
