+++ src/sfox/sfox.js
@@ -32,13 +32,14 @@ var API = require('./api');
 var Profile = require('./profile');
 var Trade = require('./trade');
 var Quote = require('./quote');
-var PaymentMethod = require('./payment-method');
+var PaymentMedium = require('./payment-medium');
+var PaymentAccount = require('./payment-account');
 
 var assert = require('assert');
 
 class SFOX extends Exchange {
   constructor (object, delegate) {
-    super(delegate, Trade, Quote, PaymentMethod);
+    super(delegate, Trade, Quote, PaymentMedium);
 
     var obj = object || {};
     this._partner_id = null;
